split canvas display controller to turtle and normal canvas ctx render
